feat: refactor workflow to split into clean rules for each technology and build rule to collect the data in one file per planning year
