+++ rules/build_electricity.smk
@@ -465,7 +465,7 @@ rule build_renewable_profiles_pecd:
         "../scripts/build_renewable_profiles_pecd.py"
 
 
-rule build_pemmdb_data:
+rule clean_pemmdb_capacities:
     params:
         snapshots=config_provider("snapshots"),
         drop_leap_day=config_provider("enable", "drop_leap_day"),
@@ -478,12 +478,99 @@ rule build_pemmdb_data:
         busmap=resources("busmap_base_s_all.csv"),
     output:
         pemmdb_capacities=resources("pemmdb_capacities_{tech}_{planning_horizons}.csv"),
-        pemmdb_p_min_pu=resources("pemmdb_p_min_pu_{tech}_{planning_horizons}.nc"),
     log:
-        logs("build_pemmdb_data_{tech}_{planning_horizons}.log"),
+        logs("clean_pemmdb_capacities_{tech}_{planning_horizons}.log"),
     threads: 4
     benchmark:
-        benchmarks("build_pemmdb_data_{tech}_{planning_horizons}")
+        benchmarks("clean_pemmdb_capacities_{tech}_{planning_horizons}")
+    conda:
+        "../envs/environment.yaml"
+    script:
+        "../scripts/clean_pemmdb_capacities.py"
+
+
+rule clean_pemmdb_profiles:
+    params:
+        snapshots=config_provider("snapshots"),
+        drop_leap_day=config_provider("enable", "drop_leap_day"),
+        available_years=config_provider(
+            "electricity", "pemmdb_capacities", "available_years"
+        ),
+        tyndp_scenario=config_provider("tyndp_scenario"),
+    input:
+        pemmdb_dir="data/tyndp_2024_bundle/PEMMDB2",
+        busmap=resources("busmap_base_s_all.csv"),
+    output:
+        pemmdb_profiles=resources("pemmdb_profiles_{tech}_{planning_horizons}.nc"),
+    log:
+        logs("clean_pemmdb_profiles_{tech}_{planning_horizons}.log"),
+    threads: 4
+    benchmark:
+        benchmarks("clean_pemmdb_profiles_{tech}_{planning_horizons}")
+    conda:
+        "../envs/environment.yaml"
+    script:
+        "../scripts/clean_pemmdb_profiles.py"
+
+
+pemmdb_techs = branch(
+    config_provider("electricity", "pemmdb_capacities", "enable"),
+    config_provider("electricity", "pemmdb_capacities", "technologies"),
+)
+
+
+def enabled_pemmdb_techs(w):
+    tyndp_conventional_carriers = config_provider(
+        "electricity", "tyndp_conventional_carriers"
+    )(w)
+    tyndp_renewable_carriers = config_provider(
+        "electricity", "tyndp_renewable_carriers"
+    )(w)
+    pemmdb_techs_dict = pemmdb_techs(w)
+
+    if not pemmdb_techs_dict:
+        return []
+
+    carrier_set = set(tyndp_conventional_carriers)
+    return [
+        tech
+        for tech, values in pemmdb_techs_dict.items()
+        if carrier_set.intersection(values)
+    ]
+
+
+def input_pemmdb_data(w):
+    enabled_techs = enabled_pemmdb_techs(w)
+    capacities = {
+        f"pemmdb_capacities_{tech}": resources(
+            f"pemmdb_capacities_" + tech + "_{planning_horizons}.csv"
+        )
+        for tech in enabled_techs
+    }
+
+    profiles = {
+        f"pemmdb_profiles_{tech}": resources(
+            f"pemmdb_profiles_" + tech + "_{planning_horizons}.nc"
+        )
+        for tech in enabled_techs
+    }
+
+    return {**capacities, **profiles}
+
+
+rule build_pemmdb_data:
+    params:
+        pemmdb_techs=enabled_pemmdb_techs,
+    input:
+        unpack(input_pemmdb_data),
+    output:
+        pemmdb_capacities=resources("pemmdb_capacities_{planning_horizons}.csv"),
+        pemmdb_profiles=resources("pemmdb_profiles_{planning_horizons}.nc"),
+    log:
+        logs("build_pemmdb_data_{planning_horizons}.log"),
+    threads: 4
+    benchmark:
+        benchmarks("build_pemmdb_data_{planning_horizons}")
     conda:
         "../envs/environment.yaml"
     script:
