+++ rules/build_sector.smk
@@ -1386,39 +1386,6 @@ def input_offshore_hubs(w):
     return {}
 
 
-pemmdb_techs = branch(
-    config_provider("electricity", "pemmdb_capacities", "enable"),
-    config_provider("electricity", "pemmdb_capacities", "technologies"),
-)
-
-
-def input_pemmdb_data(w):
-    tyndp_conventional_carriers = config_provider(
-        "electricity", "tyndp_conventional_carriers"
-    )(w)
-    tyndp_renewable_carriers = config_provider(
-        "electricity", "tyndp_renewable_carriers"
-    )(w)
-    pemmdb_techs_dict = pemmdb_techs(w)
-
-    if not pemmdb_techs_dict:
-        return {}
-
-    carrier_set = set(tyndp_conventional_carriers)
-    enabled_techs = [
-        tech
-        for tech, values in pemmdb_techs_dict.items()
-        if carrier_set.intersection(values)
-    ]
-
-    return {
-        f"pemmdb_capacities_{tech}": resources(
-            f"pemmdb_capacities_" + tech + "_{planning_horizons}.csv"
-        )
-        for tech in enabled_techs
-    }
-
-
 rule prepare_sector_network:
     params:
         time_resolution=config_provider("clustering", "temporal", "resolution_sector"),
@@ -1461,7 +1428,6 @@ rule prepare_sector_network:
         unpack(input_profile_pecd),
         unpack(input_heat_source_power),
         unpack(input_offshore_hubs),
-        unpack(input_pemmdb_data),
         **rules.cluster_gas_network.output,
         **rules.build_gas_input_locations.output,
         snapshot_weightings=resources(
@@ -1610,6 +1576,16 @@ rule prepare_sector_network:
         profile_pemmdb_hydro=branch(
             config_provider("electricity", "pemmdb_hydro_profiles", "enable"),
             resources("profile_pemmdb_hydro.nc"),
+            [],
+        ),
+        pemmdb_capacities=branch(
+            config_provider("electricity", "pemmdb_capacities", "enable"),
+            resources("pemmdb_capacities_{planning_horizons}.csv"),
+            [],
+        ),
+        pemmdb_profiles=branch(
+            config_provider("electricity", "pemmdb_capacities", "enable"),
+            resources("pemmdb_profiles_{planning_horizons}.nc"),
             [],
         ),
     output:
