+++ scripts/build_pemmdb_data.py
@@ -2,47 +2,24 @@
 #
 # SPDX-License-Identifier: MIT
 """
-Loads and cleans the available PEMMDB v2.4 data from TYNDP data bundle for a given
-
-* climate year,
-* planning horizon,
-* technology.
-
-Available technologies are:
-
-* Conventionals
-    * Nuclear
-    * Hard coal
-    * Lignite
-    * Gas
-    * Light oil
-    * Heavy oil
-    * Oil shale
-    * Other Non-RES
-* Renewables
-    * Wind
-    * Solar
-    * Hydro
-    * Other RES
-* Hydrogen
-    * Hydrogen fuel cell
-    * Hydrogen CCGT
-* Reserves
-* DSR
-* Battery
-* Electrolyser
-
+Collects and bundles the available PEMMDB v2.4 capacities and profiles for different PEMMDB technologies from TYNDP data bundle for a given planning horizon.
 
 Outputs
 -------
-Cleaned csv file with NT capacities (p_nom) in long format and netcdf file with must run obligations (p_min_pu) profiles.
+Cleaned csv file with all NT capacities (p_nom) in long format and netcdf file with must run obligations (p_min_pu) and availability (p_max_pu) for all different PEMMDB technologies.
+
+- ``resources/pemdb_capacties_{planning_horizon}.csv`` in long format
+- ``resources/pemmdb_profiles_{planning_horizon}.nc`` with the following structure:
+
+    ===================  ====================  =========================================================
+    Field                Dimensions            Description
+    ===================  ====================  =========================================================
+    profile              time, bus, carrier,   the per unit hourly availability and must-run obligations
+                         type                  for each bus and PEMMDB technology
+    ===================  ====================  =========================================================
 """
 
 import logging
-import multiprocessing as mp
-import os
-from functools import partial
-from pathlib import Path
 
 import pandas as pd
 import xarray as xr
@@ -50,258 +27,10 @@ from tqdm import tqdm
 
 from scripts._helpers import (
     configure_logging,
-    get_snapshots,
-    safe_pyear,
     set_scenario_config,
 )
 
 logger = logging.getLogger(__name__)
-
-CONVENTIONALS = [
-    "Nuclear",
-    "Hard coal",
-    "Lignite",
-    "Gas",
-    "Light oil",
-    "Heavy oil",
-    "Oil shale",
-]
-
-
-def read_pemmdb_capacities(
-    node: str,
-    pemmdb_dir: str,
-    cyear: str,
-    pyear: int,
-    pemmdb_tech: str,
-) -> pd.Series:
-    fn = Path(
-        pemmdb_dir,
-        str(pyear),
-        f"PEMMDB_{node.replace('GB', 'UK')}_NationalTrends_{pyear}.xlsx",
-    )
-
-    if not os.path.isfile(fn):
-        return None
-
-    # Conventionals & Hydrogen
-    if pemmdb_tech in CONVENTIONALS or pemmdb_tech == "Hydrogen":
-        # read capacities (p_nom)
-        df = (
-            pd.read_excel(
-                fn,
-                sheet_name="Thermal",
-                skiprows=7,  # first seven rows are metadata
-                usecols=[0, 1, 2],
-                names=["carrier", "type", "p_nom"],
-            )
-            .drop([0, 1, 2])
-            .dropna(how="all")
-            .assign(
-                **{
-                    "carrier": lambda df: df["carrier"].ffill(),
-                    "efficiency": 1.0,  # capacities are in MWel and no efficiencies are given
-                    "type": lambda df: df["type"].fillna(df.carrier),
-                    "bus": node,
-                    "country": node[:2],
-                }
-            )
-            .query("carrier == @pemmdb_tech")
-        )
-
-    # Other Non-RES
-    elif pemmdb_tech == "Other Non-RES":
-        df = (
-            pd.read_excel(
-                fn,
-                sheet_name="Other Non-RES",
-                skiprows=7,  # first seven rows are metadata
-                index_col=1,
-                nrows=9,  # capacity information is only at the top of the file
-            )
-            .filter(like="Price Band")
-            .set_axis(
-                [
-                    "p_nom",
-                    "units_count",
-                    "type",
-                    "purpose",
-                    "price",
-                    "efficiency",
-                    "co2_factor",
-                    "cyear_start",
-                    "cyear_end",
-                ]
-            )
-            .T.assign(
-                **{
-                    "carrier": "Other Non-RES",
-                    "bus": node,
-                    "country": node[:2],
-                    "cyear_start": lambda x: x.cyear_start.astype(int),
-                    "cyear_end": lambda x: x.cyear_end.astype(int),
-                }
-            )
-            .query("cyear_start <= @cyear and cyear_end >= @cyear")
-            .reset_index(drop=True)
-        )
-
-    # Wind
-    elif pemmdb_tech == "Wind":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Wind",
-        )
-
-    # Solar
-    elif pemmdb_tech == "Solar":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Solar",
-        )
-
-    # Hydro
-    elif pemmdb_tech == "hydro":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Hydro",
-        )
-
-    # Other RES
-    elif pemmdb_tech == "Other RES":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Other RES",
-        )
-
-    # Reserves
-    elif pemmdb_tech == "Reserves":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Reserves",
-        )
-
-    # DSR
-    elif pemmdb_tech == "DSR":
-        df = pd.read_excel(
-            fn,
-            sheet_name="DSR",
-        )
-
-    # Battery
-    elif pemmdb_tech == "Battery":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Battery",
-        )
-
-    # Electrolyser
-    elif pemmdb_tech == "Electrolyser":
-        df = pd.read_excel(
-            fn,
-            sheet_name="Electrolyser",
-        )
-
-    else:
-        return None
-
-    return df
-
-
-def read_pemmdb_must_runs(
-    node: str,
-    pemmdb_dir: str,
-    tyndp_scenario: str,
-    cyear: str,
-    pyear: int,
-    pyear_i: int,
-    pemmdb_tech: str,
-    sns: pd.DatetimeIndex,
-    index_year: pd.DatetimeIndex,
-) -> pd.Series:
-    fn = Path(
-        pemmdb_dir,
-        str(pyear),
-        f"PEMMDB_{node.replace('GB', 'UK')}_NationalTrends_{pyear}.xlsx",
-    )
-
-    if not os.path.isfile(fn):
-        return None
-
-    # Conventionals & Hydrogen
-    if pemmdb_tech in CONVENTIONALS or pemmdb_tech == "Hydrogen":
-        # read must runs
-        must_runs = (
-            pd.read_excel(
-                fn,
-                sheet_name="Thermal",
-                skiprows=7,  # first seven rows are metadata
-                usecols=[0, 1, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18],
-                names=[
-                    "carrier",
-                    "type",
-                    "unit",
-                    "Jan",
-                    "Feb",
-                    "Mar",
-                    "Apr",
-                    "May",
-                    "Jun",
-                    "Jul",
-                    "Aug",
-                    "Sep",
-                    "Oct",
-                    "Nov",
-                    "Dec",
-                ],
-            )
-            .drop([0, 1, 2])
-            .dropna(how="all")
-            .assign(
-                **{
-                    "carrier": lambda df: df.carrier.ffill(),
-                    "type": lambda df: df.type.ffill().fillna(df.carrier),
-                }
-            )
-            .query("unit == '% of installed capacity'")
-            .drop(columns=["unit"])
-            .set_index(["carrier", "type"])
-            .div(1e2)  # percentage conversion
-            .query("carrier == @pemmdb_tech")
-        )
-        must_runs = must_runs.set_index(must_runs.index.map("_".join)).T
-
-        df = pd.DataFrame({"month": index_year.month_name().str[:3]}, index=index_year)
-
-        for type in must_runs.columns:
-            df[type] = df["month"].map(must_runs[type])
-        df.drop(columns="month", inplace=True)
-
-        profile = (
-            df.reset_index(names="time")
-            .melt(id_vars="time", var_name="type", value_name="p_min_pu")
-            .assign(carrier=pemmdb_tech, bus=node)
-            .set_index(["time", "bus", "carrier", "type"])
-            .to_xarray()
-            .sel(time=sns)  # filter for snapshots only
-        )
-
-        # For DE and GA scenario, must-runs are removed after 2030 as to 2024 TYNDP Methodology report, p.37
-        if tyndp_scenario != "NT" and pyear_i > 2030:
-            profile = profile.assign(p_min_pu=0.0)
-
-    # Other RES
-    elif pemmdb_tech == "Other RES":
-        must_runs = pd.read_excel(
-            fn,
-            sheet_name="Other RES",
-        )
-
-    else:
-        return None
-
-    return profile
-
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
@@ -311,92 +40,33 @@ if __name__ == "__main__":
             "build_pemmdb_data",
             clusters="all",
             planning_horizons=2030,
-            tech="Gas",
         )
     configure_logging(snakemake)
     set_scenario_config(snakemake)
 
-    # Climate year from snapshots
-    sns = get_snapshots(snakemake.params.snapshots, snakemake.params.drop_leap_day)
-    cyear = sns[0].year
-    index_year = pd.date_range(
-        start=f"{cyear}-01-01",
-        periods=8760,  # 53 weeks
-        freq="h",
-    )
-    # Only climate years 1995, 2008 and 2009 are available for all technologies and countries
-    if int(cyear) not in [1995, 2008, 2009]:
-        logger.warning(
-            "Snapshot year doesn't match available TYNDP data. Falling back to 2009."
+    # Parameter
+    pemmdb_techs = snakemake.params.pemmdb_techs
+
+    # Capacities
+    capacities = []
+    logger.info("Load PEMMDB capacities for ...")
+    for tech in tqdm(pemmdb_techs):
+        capacity = pd.read_csv(
+            snakemake.input[f"pemmdb_capacities_{tech}"], index_col=0
         )
-        cyear = 2009
+        capacities.append(capacity)
 
-    # Planning year
-    pyear_i = int(snakemake.wildcards.planning_horizons)
-    pyear = safe_pyear(
-        pyear_i,
-        available_years=snakemake.params.available_years,
-        source="PEMMDB",
+    capacities_df = pd.concat(capacities, axis=0).reset_index(drop=True)
+    capacities_df.to_csv(snakemake.output.pemmdb_capacities)
+
+    # Profiles
+    profiles = []
+    logger.info(
+        f"Load PEMMDB must-run and availability profiles for {', '.join(pemmdb_techs)}..."
     )
+    for tech in tqdm(pemmdb_techs):
+        profile = xr.open_dataset(snakemake.input[f"pemmdb_profiles_{tech}"])
+        profiles.append(profile)
 
-    # Parameters
-    onshore_buses = pd.read_csv(snakemake.input.busmap, index_col=0)
-    nodes = onshore_buses.index
-    pemmdb_dir = snakemake.input.pemmdb_dir
-    tech = str(snakemake.wildcards.tech)
-    tyndp_scenario = snakemake.params.tyndp_scenario
-
-    # Load and prep inflow data
-    tqdm_kwargs_1 = {
-        "ascii": False,
-        "unit": " nodes",
-        "total": len(nodes),
-        "desc": "Loading PEMMDB capacities",
-    }
-
-    tqdm_kwargs_2 = {
-        "ascii": False,
-        "unit": " nodes",
-        "total": len(nodes),
-        "desc": "Loading PEMMDB must-runs",
-    }
-
-    func_caps = partial(
-        read_pemmdb_capacities,
-        pemmdb_dir=pemmdb_dir,
-        cyear=cyear,
-        pyear=pyear,
-        pemmdb_tech=tech,
-    )
-
-    func_must_runs = partial(
-        read_pemmdb_must_runs,
-        pemmdb_dir=pemmdb_dir,
-        tyndp_scenario=tyndp_scenario,
-        cyear=cyear,
-        pyear=pyear,
-        pyear_i=pyear_i,
-        pemmdb_tech=tech,
-        sns=sns,
-        index_year=index_year,
-    )
-
-    with mp.Pool(processes=snakemake.threads) as pool:
-        pemmdb_capacities = list(tqdm(pool.imap(func_caps, nodes), **tqdm_kwargs_1))
-        profiles = [
-            profile
-            for profile in tqdm(pool.imap(func_must_runs, nodes), **tqdm_kwargs_2)
-            if profile is not None
-        ]
-
-    pemmdb_capacities_df = (
-        pd.concat(pemmdb_capacities, axis=0)[
-            ["bus", "carrier", "type", "p_nom", "efficiency", "country"]
-        ]  # # select and order relevant columns
-    ).reset_index(drop=True)
-
-    # merge p_min_pu profiles into one xarray dataset
     ds = xr.merge(profiles)
-
-    pemmdb_capacities_df.to_csv(snakemake.output.pemmdb_capacities)
-    ds.to_netcdf(snakemake.output.pemmdb_p_min_pu)
+    ds.to_netcdf(snakemake.output.pemmdb_profiles)
